+++ srcs/create_empty_stack.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   create_empty_stack.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jleroux <marvin@42lausanne.ch>             +#+  +:+       +#+        */
+/*   By: jleroux <jleroux@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/28 16:40:20 by jleroux           #+#    #+#             */
-/*   Updated: 2022/06/29 10:50:08 by jleroux          ###   ########.fr       */
+/*   Updated: 2022/06/29 13:45:22 by jleroux          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ t_stack	create_empty_stack(int size)
 {
 	t_stack	stack;
 
-	stack.arr = (int *)malloc(sizeof(int) * size);
+	stack.arr = (long int *)malloc(sizeof(long int) * size);
 	stack.size = 0;
 	return (stack);
 }
